+++ lib/constants.py
@@ -1,4 +1,4 @@
-VERSION_NUMBER = '1.0.110'
+VERSION_NUMBER = '1.0.111'
 APPLANGA_HOST = 'https://api.applanga.com'
 API_BASE_PATH = '/v1/api'
 CONFIG_FILE_NAME = '.applanga.json'